do not make machine instance for each assertion
for example, when two assertions are related to one module:

```
(module ...)
(assert_return (invoke "a" ...) ...)
(assert_return (invoke "b" ...) ...)
```

both functions `a` and `b` must be called in the same execution context.
For instance, when `a` modifies memories or globals, `b` should be able
to access to them.

This is a bit tricky because `Machine` refers (borrows) `Module`.
Method like below is not available:

```rust
fn new_machine(&mut self, module: ast::Module<'s>) {
    let importer = ...;
    let machine = Machine::instantiate(&module, importer).unwrap();
    self.push((module, machine));
}
```

because tuple `(module, machine)` is a self-referential struct where
`machine` refers `module`. This does not work when more elements are
pushed to the `Vec` and reallocation happens to enlarge capacity.
If it is possible, the `module` instance is moved to new place and
`machine` will still refer the previous (dangling) `Module` instance.

This commit avoids it by giving longer lifetime to modules than machines
with parsing modules in advance.

```rust
fn new_machine<'a>(&mut self, module: &'a ast::Module<'s>) {
    let importer = ...;
    let machine = Machine::instantiate(module, importer).unwrap();
    self.push((module, machine));
}
```

where 'a is longer than '_.